// src/types/enum.ts
/**
 * @作者 阿瑞
 * @功能 系统枚举类型定义
 * @版本 1.0.0
 */

/* 本地存储键名枚举 */
export enum StorageEnum {
  Settings = 'saas-settings',
  Token = 'saas-token',
  UserInfo = 'saas-user-info'
}

/* 主题模式枚举 */
export enum ThemeMode {
  Light = 'light',
  Dark = 'dark'
} 